Add future tense artifact headers
Each of the three artifacts listed under goals (features list, use cases, mock UI) is given a new blurb in future tense, as per the expectations of this section. The survey has not been given a blurb because it was not part of the original plan. The old artifacts and their descriptions have not yet been removed.