Add product artifacts to review
All product artifacts, including the survey, have been moved into the review file. Each has a blurb that aims to describe the end results of the artifact, as opposed to its initial plans in the plan file. The product artifacts are ordered according to importance as specified in the sheet.