Rearrange process/project management sections
Several sections were duplicated between the "Goals and/or tasks that were planned but not met/completed" section under product management and the "Planned changes" section under process management. These sections were consolidated and placed in process management. Additionally, a blurb on the ticket list was added to the product management section.