Introduce model template
This first commit contains the models section of a Django database API,
equipped to allow access to an existing PairWise database. The
credentials of this database are not stored publicly on the repository,
and so the models currently do not work. Each developer should obtain
access to their own copy of the settings file to be able to connect to
the database.

The code enclosed does work properly with Django 1.9 and Python 3.6, as
class Model is missing attributes (Model.objects), and Manager is
missing methods (all(), filter(), etc). The code should be tested in
other environments in case this is an environment issue.