Upload group formation backends
API has been constructed for single users forming groups.
A user can enter a search, view search results, and form
a group with a single other user. Invites are not yet
implemented, nor is anything after group formation, such as
continuing to search and adding more members to the group.